User finished

+++ src/user/user.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { User } from './user.schema';
 import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import UserDto from './dto/user.dto';
+import { Role } from 'src/roles/role.enum';
 
 @Injectable()
 export class UserService {
@@ -15,20 +16,46 @@ export class UserService {
         return await this.userModel.findOne({ email: email });
     }
 
+    async findAll(): Promise<UserDto[]> {
+        return await this.userModel.find();
+    }
+
+    async getProfile(id: string): Promise<UserDto> {
+        const user = await this.userModel.findOne({ id: id });
+        user.password = undefined;
+        return user;
+    }
+
     async create(user: UserDto): Promise<UserDto> {
         const createdUser = new this.userModel(user);
         return await createdUser.save();
     }
 
-    async findAll(): Promise<UserDto[]> {
-        return await this.userModel.find();
+    async delete(id: string, role: number): Promise<UserDto> {
+        if (role == Role.Admin) {
+            return await this.userModel.findByIdAndRemove(id);
+        }
+        const user = await this.userModel.findById(id);
+        if (user.id == id) {
+            return await this.userModel.findByIdAndRemove(id);
+        } else {
+            throw new UnauthorizedException();
+        }
     }
 
-    async delete(id: string): Promise<UserDto> {
-        return await this.userModel.findByIdAndRemove(id);
-    }
-
-    async update(id: string, user: UserDto): Promise<UserDto> {
-        return await this.userModel.findByIdAndUpdate(id, user, { new: true });
+    async update(id: string, user: UserDto, role: number): Promise<UserDto> {
+        if (role == Role.Admin) {
+            return await this.userModel.findByIdAndUpdate(id, user, {
+                new: true,
+            });
+        }
+        const userDB = await this.userModel.findOne({ id: id });
+        if (userDB.id == id) {
+            return await this.userModel.findByIdAndUpdate(id, user, {
+                new: true,
+            });
+        } else {
+            throw new UnauthorizedException();
+        }
     }
 }
